+++ app.js
@@ -3,12 +3,12 @@ import bodyParser from 'body-parser';
 import { v4 as uuid } from 'uuid';
 import { NAMESPACES as ns } from './env';
 import { BASES as b } from './env';
-import * as hel from './helpers';
-import * as del from './deltaProcessing';
-import * as hea from './healing';
+import * as hel from './util/helpers';
+import * as del from './pipeline/deltaProcessing';
+import * as hea from './pipeline/healing';
 import * as test from './test/test';
 import * as env from './env';
-import * as pbu from './parse-bindings-utils';
+import * as pbu from './util/parse-bindings-utils';
 import * as mas from '@lblod/mu-auth-sudo';
 import * as rst from 'rdf-string-ttl';
 import * as cron from 'node-cron';
